Add: Authentication and middleware,Update: User Biography logic
