+++ src/main/java/com/SocialNetSys/NetSys/Controllers/User/BiographyController.java
@@ -1,7 +1,7 @@
 package com.SocialNetSys.NetSys.Controllers.User;
 
 import com.SocialNetSys.NetSys.Models.Objects_Model.BiographyRequest;
-import com.SocialNetSys.NetSys.Services.Users.Interfaces.ICreateUserBioService;
+import com.SocialNetSys.NetSys.Services.User.IUserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PatchMapping;
@@ -12,11 +12,11 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/api/v1/biography")
 public class UserBioController {
     @Autowired
-    ICreateUserBioService _CreateUSerBioService;
+    IUserService _createUserService;
 
     @PatchMapping()
     ResponseEntity<String> userBiographyCreator(@RequestBody BiographyRequest request) {
-        _CreateUSerBioService.createBiography(request);
+        _createUserService.createBiography(request);
         return ResponseEntity.ok("Created with success!");
     }
 }
