--- src/main/java/com/SocialNetSys/NetSys/Services/Users/Interfaces/ICreateUserBioService.java
@@ -1,7 +1,0 @@
-package com.SocialNetSys.NetSys.Services.Users.Interfaces;
-
-import com.SocialNetSys.NetSys.Models.Objects_Model.BiographyRequest;
-
-public interface ICreateUserBioService {
-    public String createBiography(BiographyRequest request);
-}
